Migrate `ZLibrary` calls to `BdApi` & Change class names
ZeresPluginLibrary is, currently, all but deprecated and its use is not recommended. All main functional parts of the plugin using ZPL have been swapped out for `BdApi `-equivalents for reliability and future-proofing.
While I initially planned on it, I did not fully replace ZPL because (1) `BdApi` does not have a changelog system yet for plugins and I didn't want to write my own, and (2) Your plugin settings menu also seems to be tied into ZPL and I didn't want to mess with it.

Class names have been updated to not include "vc" at the front, as this isn't a vencord plugin, even if it originates from one.

If I made any mistakes, my apologies! This is my first time contributing to somebody else's plugin. Wish you luck in getting approved!